% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_pik.R
\name{aggregate_pik}
\alias{aggregate_pik}
\title{Aggregate parsed PiK data}
\usage{
aggregate_pik(.data, .period = c("24h", "12h"), .tz = "Europe/Moscow")
}
\arguments{
\item{.data}{dataframe. Outcome of the \code{\link[=parse_pik]{parse_pik()}}
function.}

\item{.period}{character. Either \code{12h} or \code{24h} --
a string specifying the desired aggregation period.}

\item{.tz}{character. A string describing timezone of the
meteostation of interest. See \code{\link[lubridate]{with_tz}}
for details.}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
This functions calculates from raw data parsed earlier with
the \code{\link[=parse_pik]{parse_pik()}} function to a daily averages. It
took rainfall precipitation sums \code{prec} measured at 06 UTC and 18 UTC
to calculate a daily precipitation sums \code{p} (For European part of Russia).
The temperature \code{t} calculates as a daily average.
}
