% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pik.R
\name{parse_pik}
\alias{parse_pik}
\title{Parse pogodaiklimat.ru}
\usage{
parse_pik(wmo_id = "27524", start_date = "2022-04-01", end_date = "2022-05-31")
}
\arguments{
\item{wmo_id}{character. WMO meteostation ID to parse}

\item{start_date}{character. The beginning of the parsing
period. Write date in the following format \code{YYYY-MM-DD}}

\item{end_date}{character. The end of the parsing period
period. Write date in the following format \code{YYYY-MM-DD}}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
A function to parse \href{http://www.pogodaiklimat.ru/}{http://www.pogodaiklimat.ru/}
3-hourly meteo data.
}
\details{
The function returns a \code{\link[tibble]{tibble}} with
following variables:

\describe{
 \item{wmo}{character. WMO index of the meteostation}
 \item{datetime_utc}{POSIXct. Date and Time of the measurement at UTC}
 \item{ta}{numeric. Air temperature at 2m above the surface, \eqn{^{\circ}C}}
 \item{td}{numeric. Dew point, \eqn{^{\circ}C}}
 \item{rh}{numeric. Relative humidity at 2m above the surface, \eqn{\%}}
 \item{ps}{numeric. Atmosphere pressure at meteostation, \eqn{hPa}}
 \item{psl}{numeric. Atmosphere pressure adjusted to the height of mean sea level, \eqn{hPa}}
 \item{prec}{numeric. \bold{Cumulative} precipitation for the last 12 hours, \eqn{mm}}
 \item{windd}{integer. Wind direction, \eqn{^{\circ}deg}}
 \item{winds_mean}{numeric. Average 10-min wind speed, \eqn{m {\cdot} s^{-1}}}
 \item{winds_max}{numeric. Maximum wind speed, \eqn{m {\cdot} s^{-1}}}
}
}
\examples{
library(rp5pik)

example <-
  parse_pik(
    wmo_id = c("20069", "27524"),
    start_date = "2022-05-01",
    end_date = "2022-05-31"
  )

example

}
