% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp5pik-package.R
\docType{package}
\name{rp5pik-package}
\alias{rp5pik}
\alias{rp5pik-package}
\title{rp5pik: Download And Preprocess Russian Meteorological Data}
\description{
This package provides a set of functions to download and preprocess meteorological data from http://www.pogodaiklimat.ru/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atsyplenkov/rp5pik}
  \item Report bugs at \url{https://github.com/atsyplenkov/rp5pik/issues}
}

}
\author{
\strong{Maintainer}: Anatoly Tsyplenkov \email{atsyplenkov@gmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID})

}
\keyword{internal}
