% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_aggregate_pik.R
\name{rp_aggregate_pik}
\alias{rp_aggregate_pik}
\title{Aggregate parsed PiK data}
\usage{
rp_aggregate_pik(
  .data,
  .period = c("24h", "12h"),
  .tz = "Europe/Moscow",
  .direction = c("center", "left")
)
}
\arguments{
\item{.data}{dataframe. Outcome of the \code{\link[=rp_parse_pik]{rp_parse_pik()}}
function.}

\item{.period}{character. Either \verb{12h} or \verb{24h} --
a string specifying the desired aggregation period.}

\item{.tz}{character. A string describing timezone of the
meteostation of interest. See \code{\link[lubridate:with_tz]{lubridate::with_tz()}}
for details.}

\item{.direction}{character. Either \code{center} or \code{left}.
The aggregation direction. Applicable only if \code{.period} is
set to \verb{12h}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
This functions calculates from raw data parsed earlier with
the \code{\link[=rp_parse_pik]{rp_parse_pik()}} function to a daily averages. It
took rainfall precipitation sums \code{prec} measured at 06 UTC and 18 UTC
to calculate a daily precipitation sums \code{p} (For European part of Russia).
The temperature \code{t} calculates as a daily average.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
