% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_get_temp50.R
\name{rp_get_temp50}
\alias{rp_get_temp50}
\title{Get 50\% rain-snow air temperature threshold}
\usage{
rp_get_temp50(poi)
}
\arguments{
\item{poi}{SpatVector. A point layer for the Point Of Interest}
}
\value{
A data.frame
}
\description{
This function uses 50\% rain-snow air temperature threshold map
created by \emph{Jennings et al.} (\href{https://www.nature.com/articles/s41467-018-03629-7}{2018})
and returns the expected threshold value in a point. Works
only for Northern Hemisphere. See \emph{Jennings et al.} (\href{https://www.nature.com/articles/s41467-018-03629-7}{2018})
for coverage details.

At this temperature precipitation occurs as rain and snow
with equal frequency, while above the threshold precipitation
is primarily rain and below primarily snow.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\references{
Jennings, K.S., Winchell, T.S., Livneh, B. et al. Spatial variation of the rain–snow temperature threshold across the Northern Hemisphere. Nat Commun 9, 1148 (2018). https://doi.org/10.1038/s41467-018-03629-7
}
