% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pik.R
\name{parse_pik}
\alias{parse_pik}
\title{Parse pogodaiklimat.ru}
\usage{
parse_pik(wmo_id = "27524", start_date = "2022-04-01", end_date = "2022-05-31")
}
\arguments{
\item{wmo_id}{character. WMO meteostation ID to parse}

\item{start_date}{character. The beginning of the parsing
period. Write date in the following format \code{YYYY-MM-DD}}

\item{end_date}{character. The end of the parsing period
period. Write date in the following format \code{YYYY-MM-DD}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
A function to parse \url{http://www.pogodaiklimat.ru/}
3-hourly meteo data.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The function returns a \code{\link[tibble:tibble]{tibble::tibble()}} with
following variables:
\itemize{
\item \code{wmo} -- character. WMO index of the meteostation 3째C
\item \code{datetime_utc} -- POSIXct. Date and Time of the measurement at UTC
\item \code{ta} -- numeric. Air temperature at 2m above the surface, 째C
\item \code{td} -- numeric. Dew point, 째C
\item \code{rh} -- numeric. Relative humidity at 2m above the surface, \%
\item \code{ps} -- numeric. Atmosphere pressure at meteostation, hPa
\item \code{psl} -- numeric. Atmosphere pressure adjusted to the height of mean sea level, hPa
\item \code{prec} -- numeric. \strong{Cumulative} precipitation for the last \strong{12 hours}, mm
\item \code{windd} -- integer. Wind direction, 째C
\item \code{winds_mean} -- numeric. Average 10-min wind speed, m/s
\item \code{winds_max} -- numeric. Maximum wind speed, m/s
}
}
\examples{
library(rp5pik)

example <-
  parse_pik(
    wmo_id = c("20069", "27524"),
    start_date = "2022-05-01",
    end_date = "2022-05-31"
  )

example

}
